+++ packages/integration-sdk-testing/src/index.ts
@@ -1,4 +1,5 @@
 export * from './context';
+export * from './config';
 export * from './executeStepWithDependencies';
 export * from './logger';
 export * from './recording';