+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 
 ### Added
 
+- Added `StepTestConfig` and `StepTestConfigWithCredentials` exports from
+  `@jupiterone/integration-sdk-testing`
 - Updated jest matchers in the following way:
 
   - added optional `_type` argument to `.toMatchGraphObjectSchema` matcher